refactor: code deduplication
